+++ src/wam2layers/tracking/backtrack.py
@@ -68,6 +68,14 @@ def backtrack(
         + tagging_mask * precip * s_lower / (s_upper + s_lower)
         - evap * s_track_relative_lower
     )
+    s_track_negative_lower = np.where(
+        s_track_lower < 0, s_track_lower / S1["s_lower"], 0
+    )
+    if np.any(s_track_negative_lower < -1e-5):
+        logger.warn(
+            f"""Negative values encountered in s_track_lower. . Check the gains output variable for details."""
+        )
+    s_track_lower = np.maximum(s_track_lower, 0)
 
     s_track_upper += (
         +horizontal_advection(s_track_relative_upper, -fx_upper, -fy_upper, bc)
@@ -77,31 +85,46 @@ def backtrack(
         )
         + tagging_mask * precip * s_upper / (s_upper + s_lower)
     )
+    s_track_negative_upper = np.where(
+        s_track_upper < 0, s_track_upper / S1["s_upper"], 0
+    )
+    if np.any(s_track_negative_upper < -1e-5):
+        logger.warn(
+            f"""Negative values encountered in s_track_upper. Check the gains output variable for details."""
+        )
+    s_track_upper = np.maximum(s_track_upper, 0)
 
-    # down and top: redistribute unaccounted water that is otherwise lost from the sytem
+    # account for negative storages that are set to zero: "numerically gained water"
+    gains = np.abs(s_track_negative_lower + s_track_negative_upper)
+
+    # lower and upper: redistribute unaccounted water that is otherwise lost from the sytem
     # TODO build in logging for lost moisture
-    lower_to_upper = np.maximum(0, s_track_lower - S0["s_lower"])
-    upper_to_lower = np.maximum(0, s_track_upper - S0["s_upper"])
-    s_track_lower = np.minimum(
-        s_track_lower - lower_to_upper + upper_to_lower, S0["s_lower"]
-    )
-    s_track_upper = np.minimum(
-        s_track_upper - upper_to_lower + lower_to_upper, S0["s_upper"]
-    )
+    overshoot_lower = np.maximum(0, s_track_lower - S1["s_lower"])
+    overshoot_upper = np.maximum(0, s_track_upper - S1["s_upper"])
+    s_track_lower = s_track_lower - overshoot_lower + overshoot_upper
+    s_track_upper = s_track_upper - overshoot_upper + overshoot_lower
+    # at this point any of the storages could still be overfull, thus stabilize and assigns losses:
+    losses_lower = np.maximum(0, s_track_lower - S1["s_lower"])
+    losses_upper = np.maximum(0, s_track_upper - S1["s_upper"])
+    losses = losses_lower + losses_upper
+    s_track_lower = np.minimum(s_track_lower, S1["s_lower"])
+    s_track_upper = np.minimum(s_track_upper, S1["s_upper"])
 
     # Update output fields
     output["e_track"] += evap * s_track_relative_lower
+    output["losses"] += losses
+    output["gains"] += gains
 
     # Bookkeep boundary losses as "tracked moisture at grid edges"
-    output["e_track"][0, :] += (s_track_upper + s_track_lower)[0, :]
-    output["e_track"][-1, :] += (s_track_upper + s_track_lower)[-1, :]
+    output["losses"][0, :] += (s_track_upper + s_track_lower)[0, :]
+    output["losses"][-1, :] += (s_track_upper + s_track_lower)[-1, :]
     s_track_upper[0, :] = 0
     s_track_upper[-1, :] = 0
     s_track_lower[0, :] = 0
     s_track_lower[-1, :] = 0
     if config.periodic_boundary is False:  # bookkeep west and east losses
-        output["e_track"][:, 0] += (s_track_upper + s_track_lower)[:, 0]
-        output["e_track"][:, -1] += (s_track_upper + s_track_lower)[:, -1]
+        output["losses"][:, 0] += (s_track_upper + s_track_lower)[:, 0]
+        output["losses"][:, -1] += (s_track_upper + s_track_lower)[:, -1]
         s_track_upper[:, 0] = 0
         s_track_upper[:, -1] = 0
         s_track_lower[:, 0] = 0
@@ -128,6 +151,8 @@ def initialize(config_file):
             "s_track_lower_restart": xr.DataArray(0, coords=grid).astype("float32"),
             "e_track": xr.DataArray(0, coords=grid).astype("float32"),
             "tagged_precip": xr.DataArray(0, coords=grid).astype("float32"),
+            "losses": xr.DataArray(0, coords=grid).astype("float32"),
+            "gains": xr.DataArray(0, coords=grid).astype("float32"),
         }
     )
 
@@ -203,6 +228,6 @@ def run_experiment(config_file):
             progress_tracker.store_intermediate_states(output)
             write_output(output, t1, config, mode="backtrack")
             # Flush previous outputs
-            output[["e_track", "tagged_precip"]] *= 0
+            output[["e_track", "tagged_precip", "losses", "gains"]] *= 0
 
     logger.info("Experiment complete.")