+++ src/wam2layers/utils/profiling.py
@@ -1,10 +1,10 @@
 import logging
 import time
 from datetime import datetime
-import xarray as xr
 
 import numpy as np
 import psutil
+import xarray as xr
 
 from wam2layers.tracking.io import load_tagging_region
 
@@ -34,6 +34,8 @@ class ProgressTracker:
         self.mode = mode
         self.total_tagged_moisture = 0
         self.tracked = 0
+        self.lost_water = 0
+        self.gained_water = 0
         self.store_intermediate_states(output)
         self.profile = Profiler()
         self.stability_correction_previous_grid = 0
@@ -52,6 +54,8 @@ class ProgressTracker:
         else:  # mode is forwardtrack
             self.tracked += totals["p_track_lower"] + totals["p_track_upper"]
             self.total_tagged_moisture += totals["tagged_evap"]
+        self.gained_water += totals["gains"]
+        self.lost_water += totals["losses"]
 
     def print_progress(self, t, output):
         """Print some useful runtime diagnostics."""
@@ -66,20 +70,30 @@ class ProgressTracker:
         still_in_atmosphere = (
             totals["s_track_upper_restart"] + totals["s_track_lower_restart"]
         )
-        # total_tracked_moisture = tracked + still_in_atmosphere
-        # tracked_percentage = tracked / total_tagged_moisture * 100
-        # lost_percentage = (1 - total_tracked_moisture / total_tagged_moisture) * 100
+        lost_water = self.lost_water + totals["losses"]
+        gained_water = self.gained_water + totals["gains"]
+
         tracked_percentage = tracked / total_tagged_moisture * 100
         in_atmos_percentage = still_in_atmosphere / total_tagged_moisture * 100
-        lost_percentage = 100 - tracked_percentage - in_atmos_percentage
+        lost_percentage = lost_water / total_tagged_moisture * 100
+        gained_percentage = gained_water / total_tagged_moisture * 100
+        closure_check_percentage = (
+            tracked_percentage
+            + in_atmos_percentage
+            + lost_percentage
+            - gained_percentage
+        )
 
         time, memory = self.profile()
 
+        # TODO: print boundary losses and internal losses separately
         logger.info(
             f"{t} - "
             f"Tracked moisture: {tracked_percentage.item():.2f}%. "
-            f"Moisture in atmosphere: {in_atmos_percentage.item():.2f}%. "
+            f"Tagged in atmosphere: {in_atmos_percentage.item():.2f}%. "
             f"Lost moisture: {lost_percentage.item():.2f}%. "
+            f"Gained moisture: {gained_percentage.item():.2f}%. "
+            f"Tagging closure: {closure_check_percentage.item():.2f}%. "
             f"Time since start: {time}s, RAM: {memory:.2f} MB"
         )