+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add information about losses to profiling and screen messages | losses and gains as output data ([#305](https://github.com/WAM2layers/WAM2layers/pull/305)).
 - Add forward tracking possibility including small bugfix in backtrack ([#289](https://github.com/WAM2layers/WAM2layers/pull/289)).
 - Add tests for preprocess and backtrack workflow ([#265](https://github.com/WAM2layers/WAM2layers/pull/265)).
 - Add tests for visualize workflow ([#271](https://github.com/WAM2layers/WAM2layers/pull/271)).